requirements.txt
asgiref==3.7.2
chardet==5.2.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0
django-extensions==3.2.3
django-import-export==3.3.4
et-xmlfile==1.1.0
lml==0.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
Pillow==10.1.0
plum-dispatch==1.7.4
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-webio==0.1.4
PyYAML==6.0.1
Spire.Xls==13.9.0
sqlparse==0.4.4
tablib==3.5.0
texttable==1.7.0
tzdata==2023.3
xlrd==2.0.1
xlwt==1.3.0
